pyproject.toml
[tool.poetry]
name = "clean-domain"
version = "0.1.0"
description = "Demo for clean project setup"
authors = ["Cornel"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "*"


[tool.poetry.group.dev.dependencies]
pytest = "*"
mypy = "*"
ruff = "*"
factory-boy = "*"

[tool.mypy]
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.ruff]
src = ["blog"]
line-length = 100

[tool.ruff.lint]
select = [
    "I",  # isort
    "F",  # flake8
    "PL",  # pylint
    "ARG",  # flake8-unused-arguments
]

[tool.ruff.format]
indent-style = "space"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
